[feature] add spinner and result text
