+++ Picroup-iOS/Routers/Main/MainTabBarController/MainTabBarController.swift
@@ -20,7 +20,6 @@ final class MainTabBarController: UITabBarController {
         self.setup()
     }
     
-    
     required init?(coder aDecoder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
