+++ Picroup-iOS/States/MeStateObject.swift
@@ -235,7 +235,7 @@ final class MeStateStore {
     func myMediaItems() -> Driver<[MediumObject]> {
         guard let items = _state.myMedia?.items else { return .empty() }
         return Observable.collection(from: items)
-            .delaySubscription(0.3, scheduler: MainScheduler.instance)
+//            .delaySubscription(0.3, scheduler: MainScheduler.instance)
             .asDriver(onErrorDriveWith: .empty())
             .map { $0.toArray() }
     }
@@ -243,7 +243,7 @@ final class MeStateStore {
     func myStaredMediaItems() -> Driver<[MediumObject]> {
         guard let items = _state.myStaredMedia?.items else { return .empty() }
         return Observable.collection(from: items)
-            .delaySubscription(0.3, scheduler: MainScheduler.instance)
+//            .delaySubscription(0.3, scheduler: MainScheduler.instance)
             .asDriver(onErrorDriveWith: .empty())
             .map { $0.toArray() }
     }
