+++ Picroup-iOS/States/RankStateObject.swift
@@ -143,7 +143,7 @@ final class RankStateStore {
     func rankMediaItems() -> Driver<[MediumObject]> {
         guard let items = _state.rankMedia?.items else { return .empty() }
         return Observable.collection(from: items)
-            .delaySubscription(0.3, scheduler: MainScheduler.instance)
+//            .delaySubscription(0.3, scheduler: MainScheduler.instance)
             .asDriver(onErrorDriveWith: .empty())
             .map { $0.toArray() }
     }
