+++ Picroup-iOS/States/UserFollowingsStateObject.swift
@@ -210,7 +210,7 @@ final class UserFollowingsStateStore {
     func userFollowingsItems() -> Driver<[UserObject]> {
         guard let items = _state.userFollowings?.items else { return .empty() }
         return Observable.collection(from: items)
-            .delaySubscription(0.3, scheduler: MainScheduler.instance)
+//            .delaySubscription(0.3, scheduler: MainScheduler.instance)
             .asDriver(onErrorDriveWith: .empty())
             .map { $0.toArray() }
     }
