+++ Picroup-iOS/SupportingFiles/Assets.xcassets/baseline_account_circle_black_24pt.imageset/Contents.json
@@ -1,0 +1,24 @@
+{
+    "images": [
+        {
+            "filename": "baseline_account_circle_black_24pt_1x.png",
+            "idiom": "universal",
+            "scale": "1x"
+        },
+        {
+            "filename": "baseline_account_circle_black_24pt_2x.png",
+            "idiom": "universal",
+            "scale": "2x"
+        },
+        {
+            "filename": "baseline_account_circle_black_24pt_3x.png",
+            "idiom": "universal",
+            "scale": "3x"
+        }
+    ],
+    "info": {
+        "author": "xcode",
+        "template-rendering-intent": "template",
+        "version": 1
+    }
+}
