+++ Picroup-iOS/UseCases/UIKit/CollectionLoadFooterView.swift
@@ -1,0 +1,13 @@
+//
+//  CollectionLoadFooterView.swift
+//  Picroup-iOS
+//
+//  Created by luojie on 2018/5/29.
+//  Copyright © 2018年 luojie. All rights reserved.
+//
+
+import UIKit
+
+final class CollectionLoadFooterView: RxCollectionReusableView {
+    @IBOutlet weak var contentView: LoadFooterView!
+}
